EVM Kernel: “Freeze” kernel migrations prior to V34
* What

This patch introduces a clear directive within the kernel's migration codebase
to explicitly freeze all migrations introduced before `StorageVersion::V34`. It
advertizes to engineers that any future development for new kernel versions
should not inadvertently alter or re-evaluate migrations up to and excluding
version 34.

* Why

We are introducing the concept of minor releases in the EVM kernel development.
Similar to security upgrades, minor releases will be crafted by backporting
patches to release branches. The first revision of Dionysus (Etherlink 4.1) is
demonstrating that this approach does not work well with our migration system,
which assumes that kernel upgrade proposals are always cut from `master`.

While not modifying existing migrations already merged in `master` remains the
preferred policy, we also add a comment to advertize clearly the migration
already deployed on Mainnet. Rewriting migrations past the comment would
effectively introduce dead code, as these migrations will never be replayed.

* How

The freeze is “implemented” by adding a prominent comment block in the
migration logic, specifically before the `StorageVersion::V34` entry (as
Dionysus’ storage version is `V33`). This comment serves as an explicit warning
to engineers, mandating that all prior migration logic must remain untouched.
Future migrations will need to be added beyond this frozen point, ensuring
backward compatibility and preventing regressions on the Mainnet.

We also need to adapt our playbooks to make sure this comment remains
up-to-date with what is deployed on Mainnet.