EVM Kernel: Drop a no-op migration
* What

This commit removes a no-op migration related to REVM from the EVM kernel.

* Why

The migration was intended to prepare for REVM enabling, but it was found to be
unnecessary as by default we don’t rewrite migrations, we add as many as we
need.

Dropping V34, while not strictly necessary, allows to have a smaller patch for
Etherlink 4.1 (Dionysus first revision).

* How

The `V34` migration in `kernel/src/migration.rs` is removed. Consequently, the
`V35` migration is renumbered to `V34`, and `V36` to `V35`. The
`StorageVersion` enum and `STORAGE_VERSION` constant are updated accordingly.
Additionally, related version checks in `bin_node/lib_dev/gas_price.ml` and
`bin_node/lib_dev/validate.ml` are adjusted to reflect the new numbering.