EVM Node: Always provide a default sequencer config
* What

This commit refactors the node configuration to make the `sequencer`
configuration always present within the main configuration type (with default
values). Previously, the entire `sequencer` configuration was an optional
field. Additionally, the specific secret key used by the sequencer, previously
a mandatory field within the `sequencer` configuration, is now optional.

* Why

This change enables a more robust and predictable default setup for
sequencer-related parameters. By making the `sequencer` configuration always
available, other components can rely on its presence, simplifying access to
fields like `time_between_blocks` or `max_number_of_chunks` even when the node
is not explicitly running in sequencer mode or a key is not provided. This
is especially useful to configure a sandbox more precisely from the
configuration file without having to provide a fake key.

* How

The `sequencer` field in the `Configuration.t` type was changed from an
`option` to a direct `sequencer` type. The `sequencer` field within the
`sequencer` record type was changed to an `option` of `Client_keys.sk_uri`. The
`sequencer_config_dft` function was updated to support an optional sequencer
key. Corresponding changes were made to the encoding and decoding logic to
reflect these type changes, and the `sequencer_config_exn` function was
narrowed to `sequencer_key_exn` to specifically extract the optional key. RPC
services were also updated with an `is_sequencer` flag to adapt behavior based
on the node's role.