EVM Node: Add GCP KMS configuration
* What

This commit introduces a new `gcp_kms` configuration type that is currently
limited to specifying the connection pool size for Google Cloud KMS. It extends
the main node configuration to include this new setting, enabling detailed
control over how the node interacts with the KMS service for signing
operations. The changes ensure that the configuration is propagated correctly
throughout the application where KMS is used.

* Why

The addition of a configurable GCP KMS connection pool is crucial for managing
the performance and resource utilization when using external key management.
This feature provides the necessary flexibility to tune the number of
concurrent connections, which can significantly impact the efficiency and
reliability of signing operations, especially under high load. It also
standardizes how KMS-specific settings are handled within the node's
configuration. Besides, this will also be leveraged when we configure how to
authenticate the node with GCP.

* How

A `gcp_kms` record type with a `pool_size` field was defined in the
`Configuration` module, along with its encoding and a default value. The
`Configuration.t` type was extended to incorporate this new `gcp_kms` field.
Consequently, functions in the `Signer` module that interact with GCP KMS, such
as `of_sequencer_key` and `of_string`, were updated to accept the full
`Configuration.t` object, allowing them to access the configured pool size. The
`main` executable's command-line parsing and `start_sequencer` function were
adapted to correctly retrieve and pass this configuration to the relevant
`Signer` functions. The `public_key` field was also removed from
`sandbox_config` as it became redundant.