Rollup-node: Log Etherlink block numbers and evaluation time
This commit introduces logging for the time it takes to evaluate an
Etherlink block within the smart rollup node.

A new event, `eval_etherlink_block`, is added to log the block number
and the evaluation time. This is achieved by recording a timestamp
before the PVM evaluation and calculating the elapsed time after the
kernel has signaled that the block has been stored.

This provides valuable performance metrics for the Etherlink kernel.