EVM Node: Use the same URI prefix for GCP KMS as Google
* What

This commit corrects the expected prefix for Google Cloud Key Management
Service (GCP KMS) URIs within the configuration module, aligning us with the
convention implemented in the [gcpkms Go package].

[gcpkms Go package]: https://pkg.go.dev/gocloud.dev/secrets/gcpkms#pkg-constants

* Why

The previous implementation used "gcp-kms://" as the expected prefix, which is
inconsistent with the gcpkms Go package. Between us and Google, I am putting my
money on Google for being right on this one.

* How

The change involves updating two occurrences of the URI prefix: one in the
`String.remove_prefix` function, which is used for parsing an incoming key
handler string, and another in the `Format.sprintf` function, which is
responsible for formatting the URI when encoding a GCP key. Both instances now
use "gcpkms://" to ensure uniformity.