Etherlink/Tests: add __FILE__ argument to register_all
In 5dd05c87c9d94241e81a012687c79b967d2e4061, I moved the
`register_all` function to the setup.ml file, which changes the result
of the __FILE__ argument it passes to Tezt register functions and
leads to all the tests registered with register_all to be considered
by Tezt as belonging to setup.ml. This commit fixes the issue by
adding a __FILE__ argument to register_all.