EVM Node: Refactor kernel execution to allow skipping stage one
* What

This patch refactors the interface for executing the EVM kernel. It introduces
a new type for kernel inputs that explicitly distinguishes between processing
an inbox of messages and skipping the inbox-processing stage (stage one)
entirely.

* Why

The EVM kernel's execution is divided into stages. Stage one is responsible for
parsing and processing messages from the inbox. While we are now injecting the
blueprint chunks directly in the durable storage, we are still fruitlessly
running the inbox-processing stage. As a consequence, this change prevents an
unnecessary reboot.

* How

A new variant type, Wasm_runtime.kernel_input, is defined with two
constructors: `Inbox`, which wraps a list of messages, and `Skip_stage_one`.
The `Evm_state.execute` function and the Wasm runtime are updated to accept
this new type. Call sites that previously passed an empty list for block
production are modified to use `Skip_stage_one`. Other calls that handle
transactions, such as simulations or delayed transaction injections, now wrap
their messages in the `Inbox` constructor. This makes the intent of each call
to the PVM explicit and enables the kernel to skip unnecessary work.