EVM Node: Replicate the kernel checks on unsigned blueprint chunks
* What

This commit introduces a validation step in the EVM node to align the checks
performed by the `Evm_context` worker on the one performed by the kernel.

* Why

The kernel already enforces some checks on blueprint chunks (currently, only
that the maximum number of chunks of the blueprint does not exceed 128). The
EVM node, however, was not performing this check, potentially leading it to try
and apply a blueprint that the kernel would have already deemed invalid. Adding
this check ensures the EVM node fails and maintains consistency with the
kernel's validation rules.

* How

The `Apply_blueprint` handler in `evm_context.ml` is modified to
include a preliminary check on the list of chunks. If the number of
chunks exceeds the hardcoded limit of 128, the operation fails with
an error. A comment is also added in the kernel's `parsing.rs` file
to emphasize that validation logic must be replicated in the EVM
node.