Release: Octez EVM Node v0.35
This is a bug fix release that addresses two known issues in the latest
version.

First, the node now correctly handles receiving duplicated sequencer upgrade
events. There are several scenarios in which this can occur, so this fix is
necessary to ensure smooth transitions for the sequencer operator. Second, the
node will no longer format addresses in its RPC responses using the EIP-55
checksum format. This change, introduced in the previous version, was not
consistent with the behavior of Ethereum nodes and has therefore been reverted.

We strongly recommend that node providers upgrade to this version.

This release includes a single migration to the internal store (version 22),
which means it is not possible to downgrade to the previous version.