Tezt: Fix `Baking nonce format migration` ill-formated nonces file
* What

Tweak the tezt scenario to generate a well-formated file that can be parse in
its entierety. Previously, the generated file would look like this.

    [ { "block": "BLibm5QJetZcbA427EWFp6C82AvaZ5cDfVrUyEmUiRrpdfkDJNb",
         "nonce":
           "1eb0fa2bcd291dc415c37bb10a3006656c0e3c87edfe6301e8b054639c74bc5a" },
       { "block": "BLKZ45h2CHNhj1hpd5ztJXG6T9WEeBKXLKLzd55RJVtVXLmKg2R",
         "nonce":
           "2d73a1ccd11b9b08464f66866762f0db259cefc48a2e136eef4fde4a69065d87" } ][ { "block": "BMVwjFL5ZRdB1icmcJ4tjy2Uz7edz1jcZ3m872KCqmYnVEi6Hd1",
         "nonce":
           "4aea386a659f6a4f921bce12af466312e626407754e984550a22c47dac1432d1" } ]

* Why

Due to a buggy helper, the consumers of the nonce files stop trying to parse
their contents once they have read a valid JSON object. Meaning that the new
nonces were effectively ignore during the test. Generating a proper file ensure
the binary also takes into account the new nonces.

* How

Parse the new and old nonces contents, in order to combine the two arrays into
one, and delegate to `Data_encoding` the generation of a valid JSON string.