EVM Node: Refresh token on unauthorized GCP KMS sign requests
* What

This commit introduces a retry mechanism based on a refresh strategy for sign
requests failing with HTTP error 401. It aims to improve the resilience of
interactions with GCP KMS.

* Why

We are witnessing spurious Unauthorized errors for some HTTP requests made by
the sequencer to a GCP KMS. It seems like there are several scenarios which can
lead a token to be invalidated before it is being refreshed by the node.

The node is actually lacking the necessary logic to handle this edge case,
which this commits implement.

* How

The retry logic is implemented by wrapping potentially failing GCP KMS calls
with a loop that re-attempts the operation once. On the first failure, a new
token is fetched before retrying.