Tezlink/Kernel: Introduce the OriginationSuccess in operation_result.rs (and propagate it)
We introduce a type Originated that represent that represent a KT1 encoded. The encoding of this
new struct Originated is the same as the Originated branch of the Contract enum for the kernel_sdk.
But as an originated contract can't be an Impilicit Account, we're not interested in the other case
of this enum.