Proto: avoid consensus_threshold_size direct call from constants
This value will depend on the level. Correct uses of this constant
have been commented. Most uses were in the baker, so an update to the
Validators RPC has been made. Now, the validators are regrouped by
level, and level data like the consensus threshold is also given (this
avoids some redundant information about the level being sent through
the RPC for every delegate. This level field was also never used in
the baker anyways?). This is a breaking change for the RPC btw