EVM Node: Vendor lwt_domain library for parallelism
* What

This commit introduces a vendored version of the `lwt_domain`
library into the Etherlink codebase. This new internal library, named
`octez-lwt-domain`, provides functionalities to offload computations
to a separate pool of domains (OS threads) and manage their
execution from the main Lwt thread.

* Why

The EVM node, like other Octez components, relies on Lwt for
cooperative concurrency, which runs on a single core. CPU-intensive
tasks can block the main event loop, hurting responsiveness and
performance. We were relying on `Lwt_preemptive` to delegate heavy
computations to other cores, but with OCaml global lock on the
runtime, this is not enough. By switching to `lwt_domain`, we enables
true parallelism and is a prerequisite for performance optimizations
within the EVM node.

Vendoring the library will allow us to iron it out for our needs,
prior to proposing patches upstream to be integrated and, in time,
the vendored library to be dropped.

* How

A new library `octez-lwt-domain` is added under the `etherlink/`
directory, containing the source code for `lwt_domain`. The build
system is updated to recognize this new library, adding it to the
`dune-project`, manifest files, and creating a corresponding OPAM
package.

The `domainslib` library is added as a new external dependency for the
project. The `octez-lwt-domain` library is configured to depend on
`lwt`, `lwt.unix`, and `domainslib`. Finally, the `evm_node_lib_dev`
library is updated to include `octez-lwt-domain` as a dependency,
making it available for use in the EVM node.