EVM Node: Introduce a Domain-friendly alternative to Lwt_process
* What

This commit introduces a new `Process_manager` module to handle the
creation of external processes. All process-spawning operations are now
executed in a dedicated system thread, preventing the main Lwt event loop from
being blocked without relying on the `Unix.fork` function.

* Why

`Lwt_process` relies on `Unix.fork` to spawn no processes. Not only is it
blocking, but it is also incompatible with the use of OCaml5 domains.
`Process_manager` on the contrary delegates the creation of the process to a
dedicated thread using `Lwt_domain`, and use the `Unix.open_process*` functions
which are compatible with OCaml 5 domains.

* How

A new `Process_manager` module is introduced, which provides an API similar to
`Lwt_process`. Internally, it uses a single-threaded `Lwt_domain` pool to run
the blocking `Unix.open_process_*` functions. All usages of `Lwt_process`
within the EVM node, specifically in the GCP KMS and RPC server modules, have
been replaced with calls to the new `Process_manager`. Consequently,
`Rpc.spawn_main` becomes asynchronous. The call sites in the observer and
sequencer have been updated to accommodate this change.