EVM Node: Avoid fork in snapshot download check
* What

This change modifies the program_exists utility function, which is used during
snapshot imports. The modification replaces a call to Lwt_process with a more
direct and `Unix.fork`-free implementation using `Unix.open_process` and
`Lwt_unix.waitpid`.

* Why

The original implementation used `Lwt_process` to run the which
command to check if the download helper `curl was available. `Lwt_process` is
built upon `Unix.fork`, which is unsafe to call in a multi-threaded application,
such as one using OCaml 5 domains. This could cause the node to crash when a
user initiated a snapshot import from a remote URL. This commit provides a
minimal fix for this specific issue.

* How

The program_exists function no longer uses `Lwt_process`. Instead, it now
directly spawns a process using `Unix.open_process_args_full`. It then
immediately gets the process ID and uses `Lwt_unix.waitpid` to asynchronously
wait for the command to complete. This approach successfully checks for the
program's exit status without introducing the risks associated with `Unix.fork`
in a multi-threaded environment.